requirements.txt
amqp==5.3.1
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.12.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
autopep8==1.6.0
beautifulsoup4==4.10.0
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
boto3==1.37.11
botocore==1.37.11
branca==0.8.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.0
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cssselect2==0.8.0
cycler==0.12.1
db==0.1.1
db-sqlite3==0.0.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
dj-database-url==0.5.0
Django==4.2.16
django-admin-volt==1.0.10
django-ajax-datatable==4.4.2
django-allauth==65.4.1
django-computed-property==0.3.0
django-computedfields==0.1.5
django-cors-headers==4.9.0
django-crispy-forms==1.12.0
django-debug-toolbar==4.2.0
django-environ==0.8.1
django-filter==2.4.0
django-huey==1.3.0
django-ses==3.1.0
django-tables2==2.3.4
django-tables2-column-shifter==2.0.3
django-widget-tweaks==1.4.9
django-xhtml2pdf==0.0.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
email_validator==2.2.0
filelock==3.17.0
firebase-admin==6.8.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Limiter==3.10.1
flask-swagger-ui==4.11.1
folium==0.19.5
fonttools==4.56.0
geoip2==2.9.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-oauth2-tool==0.0.3
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==20.1.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huey==2.5.3
idna==3.10
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.13.0
jmespath==1.0.1
jsonpickle==4.0.5
kiwisolver==1.4.8
kombu==5.4.2
legacy-cgi==2.6.3
limits==4.0.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.1
maxminddb==2.6.3
mdurl==0.1.2
moviepy==2.2.1
msgpack==1.1.0
numpy==2.2.5
oauth2client==4.1.3
oauthlib==3.2.0
openai==2.20.0
ordered-set==4.1.0
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pbr==6.1.1
pillow==11.1.0
platformdirs==4.3.6
plotly==5.6.0
postgres==4.0
proglog==0.1.12
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.8.0
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.1
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.9.0
pymongo==4.10.1
pyparsing==3.2.3
pypdf==5.3.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.5
python-docx==1.1.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==6.0.2
qrcode==8.0
redis==6.2.0
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.4
rsa==4.9.1
s3transfer==0.11.4
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.3.1
sqlparse==0.4.2
stevedore==5.4.1
str2bool==1.1
svglib==1.5.1
tenacity==9.0.0
tinycss2==1.4.0
toml==0.10.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
tzlocal==5.3.1
Unipath==1.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.3
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==5.3.0
wrapt==1.17.2
xhtml2pdf==0.2.17
xlwt==1.3.0
xyzservices==2025.1.0
